Imported turn based combat game
The application was written according to the technical specifications, attached below

Hello, dear applicant! We present to your attention a test task that will help evaluate your programming skills, understanding of client-server interaction and proficiency in the Unity engine. When completing a task, spend most of your time on the code itself, the architecture and the absence of bugs.
The visual component of the test task will be of minimal importance.

Key Features
Your task is to create a working prototype according to the specifications. The main problem is to write logic in such a way that it fits the client-server interaction model.
To simplify the work on the test, it is recommended to write the client and server logic in the form of separate structural elements, communication through which can be organized, for example, through the adapter interface.

Scene
The scene is a battlefield on which two conventional units are located opposite each other: the player’s unit and the enemy’s unit.
Above each of the units there should be a unit’s life bar, which will also display the amount of life in numbers.

Gameplay
Units fight among themselves, using different abilities in turn. First the player chooses an ability, then the enemy, and so on until the life of one of the units drops below zero.
After which the game should restart again. There is no need to spend time developing the AI; it will be enough if the AI ​​chooses actions randomly.
Each unit has a set of several abilities: attack, barrier, regeneration, fireball, purification. Each ability, except attack and cleanse, has a cooldown and duration,
measured in strokes. Cooldown begins after the effect ends. The ability cannot be used while on cooldown.
0) Attack – applied to the enemy. Deals 8 damage.
1) Barrier – applied to yourself, blocks a total of 5 units of damage. Duration 2 turns, reload 4 turns.
2) Regeneration - applied to yourself,
Each turn restores 2 units of life. Duration 3 turns, cooldown 5 turns.
3) Fireball – when used, it deals 5 damage to the enemy, after which an effect is applied that deals 1 damage to the enemy every turn. Duration 5 turns, cooldown 6 turns.
4) Cleansing – applied to yourself, removes the burning effect. Reload 5 turns.

Interface
Should display the player's abilities and their cooldown during turns. There should also be a battle restart button in the interface.
Icons of status effects received after using abilities should appear next to the units’ life bars. On the effect icons it is worth placing a counter for the remaining number of moves, during which the effect will last.